move : relocated station test code
